+++ testmodul/isilive_test_scenarios_factory_highlight_priority.lua
@@ -152,9 +152,6 @@ return function(test, ctx)
             BuildOrderedRoster = function()
               return {}
             end,
-            HasFullSync = function()
-              return false
-            end,
             BuildDisplayData = function()
               return {}
             end,